day1/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
aiohttp>=3.11.0
python-dotenv>=1.0.0
rich>=13.9.0